Trace to events refactor algorithms (#472)
## Summary of changes

- Reworks parameters of the algorithm functions to take an iterator
rather than flatbuffer slice. This allows the functions to be used by
the upcoming multiscale pyramid algorithm.
- Moves some structs to `state.rs`
- Add a "FIXME" to handle an error in `processing.rs` in a future PR.

## Instruction for review/testing

General code review.
Has been tested in simulated pipeline.
